Break package into multiple submodules.
